chore(onboarding): moves onboarding to components
