+++ src/v2/Components/Onboarding/Components/OnboardingFigure.tsx
