+++ src/v2/Components/Onboarding/Components/OnboardingGene.tsx
@@ -3,7 +3,7 @@ import { createFragmentContainer, graphql } from "react-relay"
 import { SystemQueryRenderer } from "v2/System/Relay/SystemQueryRenderer"
 import { OnboardingGene_gene } from "v2/__generated__/OnboardingGene_gene.graphql"
 import { OnboardingGeneQuery } from "v2/__generated__/OnboardingGeneQuery.graphql"
-import { OnboardingLoadingCollection } from "v2/Apps/Onboarding/Components/OnboardingLoadingCollection"
+import { OnboardingLoadingCollection } from "v2/Components/Onboarding/Components/OnboardingLoadingCollection"
 import { ArtworkGridItemFragmentContainer } from "v2/Components/Artwork/GridItem"
 import { Masonry } from "v2/Components/Masonry"
 import { extractNodes } from "v2/Utils/extractNodes"
