+++ src/v2/Components/Onboarding/Components/OnboardingProgress.tsx
