+++ src/v2/Components/Onboarding/Views/OnboardingArtistsOnTheRise.tsx
