+++ src/v2/Components/Onboarding/Views/OnboardingCuratedArtworks.tsx
