+++ src/v2/Components/Onboarding/Views/OnboardingFollowArtists.tsx
