+++ src/v2/Components/Onboarding/config.ts
